refactoring: extract npm-registry-name and component-id-to-package-name code to new modules
