+++ src/consumer/component/dependencies/dependency-resolver/dependencies-versions-resolver.js
@@ -8,7 +8,7 @@ import type Component from '../../../component/consumer-component';
 import logger from '../../../../logger/logger';
 import type Consumer from '../../../../consumer/consumer';
 import type { PathLinux } from '../../../../utils/path';
-import getNodeModulesPathOfComponent from '../../../../utils/component-node-modules-path';
+import getNodeModulesPathOfComponent from '../../../../utils/bit/component-node-modules-path';
 import type ComponentBitJson from '../../../bit-json';
 import Dependencies from '../dependencies';
