+++ src/links/link-generator.js
@@ -261,6 +261,9 @@ function getComponentsDependenciesLinks(
 }
 
 /**
+ * important: do not attempt to move this function into DependencyFileLinkGenerator as it should be
+ * running even when a component does not have any dependencies.
+ *
  * when using custom module resolutions, and inside a component there is a file that requires
  * another file by custom-resolved syntax, we must generate links on the imported component inside
  * node_modules.
